requirements.txt
pillow==8.4.0
tk==0.1.0
fastapi~=0.110.0
requests~=2.31.0
